# Faster unit tests: directloy operate with CMutableTransaction