Refactored `TransactionSignatureCreator` so into a templated `GenericTransactionSignatureCreator` that works with both `CMutableTransaction` and `CTransaction`.

The advantage is that now in `SignSignature`, the `MutableTransactionSignatureCreator` can now operate directly with the `CMutableTransaction` without the need to copy the data into a `CTransaction`.

Running all unit tests brings a very noticable speedup on my machine:

    48.4 sec before this change
    36.4 sec with this change
    --------
    12.0 seconds saved

running only `--run_test=transaction_tests/test_big_witness_transaction`:

    16.7 sec before this change
     5.9 sec with this change
    --------
    10.8 seconds saved

This relates to my first attempt with the const_cast hack #13202, and to the slow unit test issue #10026.