@promag 2018-05-23
Nice improvement, concept ACK.