@martinus 2018-05-27
As suggested by @sipa I have now removed the template for `MutableTransactionSignatureCreator` and used it directly in wallet.cpp.
Also I ran the format script and squashed everything.