@MarcoFalke 2018-05-29
Tagged with "refactoring" since the only change should be removing the redundant member `MutableTransactionSignatureCreator::tx`