@martinus 2018-05-30
In 6b8b63a I've fixed all the nits, tried to keep the diff minimal, and squashed it again