@fanquake 2018-05-31
master (472fe8a):
```
$ time src/test/test_bitcoin -- run_test=transaction_tests/test_big_witness_transaction
Running 1 test case...
*** No errors detected
real	0m20.634s
user	0m21.253s
sys	0m0.163s
```
```
$ time src/test/test_bitcoin
Running 291 test cases...
*** No errors detected
real	1m23.903s
user	0m56.010s
sys	0m40.136s
```

This PR (6b8b63a):
```
time src/test/test_bitcoin --run_test=transaction_tests/test_big_witness_transaction
Running 1 test case...
*** No errors detected
real	0m4.316s
user	0m4.961s
sys	0m0.157s
```

```
time src/test/test_bitcoin
Running 291 test cases...
*** No errors detected
real	1m7.374s
user	0m38.687s
sys	0m41.480s
```